Move textio.FileSystem and its registry to a separate package
